Use Rate to represent transfer rates (RIPD-201, RIPD-983):
The Ripple protocol represent transfer rates and trust line
qualities as fractions of one billion. For example, a transfer
rate of 1% is represented as 1010000000.

Previously, such rates where represented either as std::uint32_t
or std::uint64_t. Other, nominally related types, also used an
integral representation and could be unintentionally substituted.

The new Rate class addresses this by providing a simple, type
safe alternative which also helps make the code self-documenting
since arithmetic operations now can be clearly understood to
involve the scaling of an amount by a rate.

+++ src/ripple/protocol/STAmount.h
@@ -295,7 +295,7 @@ public:
 //------------------------------------------------------------------------------
 
 STAmount
-amountFromRate (std::uint64_t uRate);
+amountFromRate (std::uint32_t uRate);
 
 // VFALCO TODO The parameter type should be Quality not uint64_t
 STAmount
